+++ sources/msh_util/msh_strlen.c
@@ -6,15 +6,14 @@
 /*   By: jho <jho@student.42seoul.kr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/29 20:31:32 by jho               #+#    #+#             */
-/*   Updated: 2023/10/10 12:07:20 by jho              ###   ########.fr       */
+/*   Updated: 2023/10/26 14:59:07 by jho              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
 
-size_t	msh_strlen(char *s)
+int	msh_strlen(char *s)
 {
-	size_t	len;
+	int	len;
 
 	len = 0;
 	while (*(s + len) != '\0')
